MASSIVE CHANGES 1/
New selling system. Changes to the database include the introduction of 2 new
tables for transactions. Impacts on code are huge and are being processed.

This commit contains the new pages, numerous changes to include the
icon for selling bundles and some updates to different parts of code
following the changes on data tables.